gh-4 Add Talk instruction
Also: added auto instruction. Don't currently love
how I'm doing it... Feels like it would be beter to have
GameDataProcessor push instructions to some kind of
instruction queue and then run through and process
each one.

Still need to make the auto action to get the Dagger
after talking to Élodie. Not as simple as just callling
`take dagger` because the Take instruction requires
the item to be in the current POI. Ugh.